The radar processors has been added
