+++ src/main/java/com/a5lab/axion/domain/wizard/RadarProcessor.java
@@ -1,5 +1,5 @@
 package com.a5lab.axion.domain.wizard;
 
 public interface RadarProcessor {
-  void process();
+  void process() throws Exception;
 }
